requirements.txt
django==4.2.1 
psycopg2==2.9.9
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-cors-headers==4.4.0
djangorestframework==3.14.0
sqlparse==0.4.4
Pillow==10.1.0